--- privbayes/makefile
@@ -1,2 +1,0 @@
-all:
-		g++ -O3 --std=c++14 code/main_marginal.cpp code/methods.cpp code/noise.cpp code/printer.cpp code/table.cpp code/translator.cpp -o privBayes.bin
