--- results/0.2.0.csv
@@ -1,9 +1,0 @@
-,adult/accuracy,adult/f1,alarm/syn_likelihood,alarm/test_likelihood,asia/syn_likelihood,asia/test_likelihood,census/accuracy,census/f1,child/syn_likelihood,child/test_likelihood,covtype/accuracy,covtype/macro_f1,covtype/micro_f1,credit/accuracy,credit/f1,grid/syn_likelihood,grid/test_likelihood,gridr/syn_likelihood,gridr/test_likelihood,insurance/syn_likelihood,insurance/test_likelihood,intrusion/accuracy,intrusion/macro_f1,intrusion/micro_f1,mnist12/accuracy,mnist12/macro_f1,mnist12/micro_f1,mnist28/accuracy,mnist28/macro_f1,mnist28/micro_f1,news/r2,ring/syn_likelihood,ring/test_likelihood,timestamp
-IdentitySynthesizer,,0.669,-10.3,-10.3,-2.23,-2.24,,0.494,-12.0,-12.0,,0.652,,,0.72,-3.06,-3.06,-3.06,-3.07,-12.8,-12.9,,0.862,,0.8859999999999999,,,0.9159999999999999,,,0.14,-1.7,-1.7,
-CLBNSynthesizer,,0.33399999999999996,-12.4,-11.2,-2.44,-2.27,,0.31,-12.6,-12.3,,0.319,,,0.409,-3.68,-8.62,-3.76,-11.6,-15.2,-13.9,,0.384,,0.741,,,0.17600000000000002,,,-6.28,-1.75,-1.7,
-PrivBNSynthesizer,,0.414,-11.9,-10.9,-2.28,-2.24,,0.121,-12.3,-12.2,,0.27,,,0.185,-4.33,-21.67,-3.98,-13.88,-14.7,-13.6,,0.384,,0.11699999999999999,,,0.081,,,-4.49,-1.82,-1.71,
-MedganSynthesizer,,0.375,-10.9,-14.2,-2.81,-2.59,,0.0,-14.2,-15.4,,0.09300000000000001,,,0.0,-10.04,-62.93,-9.45,-72.0,-16.4,-16.4,,0.299,,0.091,,,0.10400000000000001,,,-8.8,-2.32,-45.16,
-VEEGANSynthesizer,,0.235,-17.7,-14.9,-8.11,-4.63,,0.094,-17.6,-17.8,,0.08199999999999999,,,0.0,-9.81,-4.79,-12.51,-4.94,-18.2,-18.1,,0.261,,0.19399999999999998,,,0.136,,,-6500000.0,-7.85,-2.92,
-TableganSynthesizer,,0.49200000000000005,-12.7,-11.5,-3.64,-2.77,,0.358,-15.0,-13.3,,0.0,,,0.182,-8.7,-4.99,-9.64,-4.7,-16.0,-14.3,,0.0,,0.1,,,0.0,,,-3.09,-6.38,-2.66,
-TVAESynthesizer,,0.626,-11.2,-10.7,-2.31,-2.27,,0.377,-12.3,-12.3,,0.433,,,0.098,-2.86,-11.26,-3.41,-3.2,-14.7,-14.2,,0.511,,0.7929999999999999,,,0.794,,,-0.2,-1.68,-1.79,
-CTGANSynthesizer,,0.601,-14.2,-12.6,-2.56,-2.31,,0.391,-13.4,-12.7,,0.324,,,0.672,-5.63,-3.69,-8.11,-4.31,-16.5,-14.8,,0.528,,0.39399999999999996,,,0.371,,,-0.43,-3.43,-2.19,
