--- results/0.2.1.csv
@@ -1,11 +1,0 @@
-,adult/accuracy,adult/f1,alarm/syn_likelihood,alarm/test_likelihood,asia/syn_likelihood,asia/test_likelihood,census/accuracy,census/f1,child/syn_likelihood,child/test_likelihood,covtype/accuracy,covtype/macro_f1,covtype/micro_f1,credit/accuracy,credit/f1,grid/syn_likelihood,grid/test_likelihood,gridr/syn_likelihood,gridr/test_likelihood,insurance/syn_likelihood,insurance/test_likelihood,intrusion/accuracy,intrusion/macro_f1,intrusion/micro_f1,mnist12/accuracy,mnist12/macro_f1,mnist12/micro_f1,mnist28/accuracy,mnist28/macro_f1,mnist28/micro_f1,news/r2,ring/syn_likelihood,ring/test_likelihood,timestamp
-CLBNSynthesizer,0.7601166666666667,0.29132585109423337,-12.38543644119261,-11.181649727926832,-2.4065885199066424,-2.2718399692766007,0.8950466666666667,0.27358008626249114,-12.636519547189994,-12.307926689840349,0.26416666666666666,0.12753506639614431,0.26416666666666666,0.6819208333333333,0.16131007132601652,-3.885928070248749,-5.274841343089261,-4.066210482168392,-10.287410604034287,-15.206133224937735,-13.917470939937324,0.8877533333333331,0.34311370434141303,0.8877533333333331,0.7206,0.7156855405119821,0.7206,0.14935,0.14269916559049736,0.14935,-61.09662317853844,-1.7505920267886637,-18.904814759240423,2020-04-12 09:41:35.096775
-CTGANSynthesizer,0.7842166666666666,0.5974197011936037,-15.625476782503652,-12.915224490214193,-3.8869031035684727,-2.4423750616134168,0.9113366666666668,0.3750530081503448,-14.044374133132287,-12.832625960011804,0.56955,0.32593335389350003,0.56955,0.9976166666666668,0.6278324954628842,-9.162882145624636,-5.066747087143494,-8.653292980523496,-5.086304003131836,-16.917861328645607,-14.961378936056466,0.9733166666666668,0.4832861331967759,0.9733166666666668,0.14468333333333333,0.12203090558035015,0.14468333333333333,0.11971666666666665,0.11129519643080237,0.11971666666666665,-0.0967127382167677,-7.056121943207441,-2.7371486744037625,2020-04-12 09:41:35.096775
-IdentitySynthesizer,0.8256083333333334,0.6616002099514467,-10.233934785738253,-10.301955284430926,-2.250748288792724,-2.241122417029465,0.9074983333333334,0.4644702794304546,-12.045682408829405,-12.041519112231095,0.7569433333333334,0.6509993111078477,0.7569433333333332,0.9939083333333332,0.5768042720939585,-3.4766624001200728,-3.5032424492771077,-3.607534449563092,-3.635514265199317,-12.850256611430414,-12.95907200045215,0.9990833333333334,0.8670476550033723,0.9990833333333334,0.8805833333333334,0.8790578371871959,0.8805833333333334,0.9074500000000002,0.9062346923885176,0.9074500000000002,0.1353284044241243,-1.7134048821174204,-1.7043594708894825,2020-04-12 09:41:35.096775
-IndependentSynthesizer,0.654425,0.17451468832786352,-18.23031525304012,-15.805229803856498,-4.933313601476521,-3.0016273214249587,0.6800716666666666,0.05605739411686009,-17.949428100386573,-16.045037769167617,0.39934,0.1224435871731518,0.39934,0.8714875000000002,0.00010281062621772051,-3.5441361455121263,-3.4699707776146935,-5.03331167511021,-4.037670004366031,-18.38149639137817,-17.56380719228361,0.55448,0.14884965304354345,0.55448,0.09956666666666668,0.09679187236230267,0.09956666666666668,0.10156666666666664,0.10019940487244998,0.10156666666666664,-0.060721752973279336,-2.4655574858905718,-1.9651172111506137,2020-04-12 09:41:35.096775
-MedganSynthesizer,0.5936833333333333,0.2761666282714071,-8.169460968146819,-13.020430279107508,-1.5194222326323434,-5.197980779624778,0.8132933333333333,0.061012177474682114,-11.198929530135699,-13.09912231080883,0.47173000000000004,0.098197880586747,0.47173000000000004,0.99825,0.0,-6.833268269910192,-84.38058716048052,-7.7474771070377315,-160.89915926586556,-13.801601086992044,-14.943326526661297,0.7594666666666666,0.2706623868796735,0.7594666666666666,0.3749,0.3145170646743483,0.3749,0.09156666666666667,0.05144973012975695,0.09156666666666667,-5.2551853860103925,-2.7695432928090806,-133.46290031962073,2020-04-12 09:41:35.096775
-TableganSynthesizer,0.79885,0.4262577542033918,-12.744195264607079,-11.565230172016301,-3.3310516736101263,-2.6844904854631224,0.9359649999999999,0.2899430368587291,-14.768574533055256,-13.241063167862706,0.0,0.0,0.0,0.998175,0.600426353460308,-6.777963983549516,-4.931756213669246,-7.080974443429973,-5.047244658319892,-16.087381028522056,-14.299551867472635,,,,0.07918333333333334,0.07066194872440286,0.07918333333333334,0.07881666666666666,0.056762438352421174,0.07881666666666666,-4.22744006074679,-3.898531300201869,-2.307308459846443,2020-04-12 09:41:35.096775
-TVAESynthesizer,0.8108166666666666,0.6301746613802992,-11.357111743987383,-10.754935335278683,-2.3533268235818023,-2.2673343868022706,0.92409,0.40664050896751697,-12.285049581782431,-12.248006692866118,0.6620566666666666,0.4331591847169729,0.6620566666666666,0.9983208333333335,0.07867086945101129,-3.3882737607271367,-5.190145619137,-3.820568826085621,-3.7246326034251536,-14.293041552910989,-13.823180492487486,0.9774333333333334,0.4614386335762677,0.9774333333333334,0.7936666666666667,0.7911782725775057,0.7936666666666667,0.7723166666666667,0.7702146846500276,0.7723166666666667,-0.4426514751777253,-1.654247003638594,-1.9330562668620004,2020-04-12 09:41:35.096775
-UniformSynthesizer,0.5002416666666666,0.3018783210281324,-18.42068074395208,-18.42068073863384,-14.301041753456893,-5.551594050449191,0.44254666666666653,0.09779765205625596,-18.40974282089297,-18.333704116138865,0.18989333333333333,0.09857246330311503,0.18989333333333333,0.5902333333333333,0.01126299780596824,-7.2940523428100885,-4.534826641153023,-7.227005516201165,-4.549560036340737,-18.42068074003849,-18.420676595334662,0.10312,0.08095202581136543,0.10312,0.09605,0.08750004190133483,0.09605,0.10401666666666666,0.09222383712083117,0.10401666666666666,-4.612017729534938,-5.343588515281946,-2.5207336976727768,2020-04-12 09:41:35.096775
-VEEGANSynthesizer,0.7187916666666667,0.16214153392906325,-18.386072879347,-18.210907275423317,-11.492286773035527,-5.952711599023645,0.7643516666666667,0.05181635667822724,-17.314283585787468,-17.681451873351854,0.21852,0.09591554596683807,0.21852,0.8762583333333334,0.16603037449105235,-8.646858295086131,-423.5732756751199,-11.458545686256631,-8.908475059016894,-18.32582658070821,-18.113418656510373,0.5064133333333333,0.1807969282858515,0.5064133333333333,0.3746833333333333,0.3495709511476633,0.3746833333333333,0.15563333333333332,0.058182612869574775,0.15563333333333332,-319605180.2536416,-16.83063406863265,-6.354959956004122,2020-04-12 09:41:35.096775
-PrivBNSynthesizer,0.7935333333333333,0.4361005523815438,-12.30822824251503,-11.144571152309572,-2.27054865275148,-2.2477659341917207,0.8810250000000002,0.18437324469817612,-12.384919843908124,-12.181347951110359,0.4907833333333334,0.2415233567590648,0.4907833333333334,0.98775,0.16564359581419338,-3.908484506897013,-7.3347240981252355,-4.08618993572303,-8.532455255491085,-14.758795882340554,-13.64782756674383,0.9020566666666666,0.3533865528603071,0.9020566666666666,,,,,,,-2.920691846838815,-1.785586399630908,-502.3862268365551,2020-04-30 16:36:33.054846
