--- scripts/config/clbn.conf
@@ -1,6 +1,0 @@
-NAME=CLBN
-CPU=10
-GPU=
-SYNTHESIZERS=CLBN
-MODALITIES=single-table
-CACHE=TRUE
