--- scripts/config/copulagan.conf
@@ -1,6 +1,0 @@
-NAME=CopulaGAN
-CPU=20
-GPU=2
-SYNTHESIZERS=CopulaGAN
-MODALITIES=single-table
-CACHE=TRUE
