--- scripts/config/copulas.conf
@@ -1,6 +1,0 @@
-NAME=Copulas
-CPU=40
-GPU=
-SYNTHESIZERS="GaussianCopulaCategorical GaussianCopulaCategoricalFuzzy GaussianCopulaOneHot"
-MODALITIES=single-table
-CACHE=TRUE
