--- scripts/config/ctgan.conf
@@ -1,6 +1,0 @@
-NAME=CTGAN
-CPU=20
-GPU=2
-SYNTHESIZERS=CTGAN
-MODALITIES=single-table
-CACHE=TRUE
