--- scripts/config/hma1.conf
@@ -1,6 +1,0 @@
-NAME=HMA1
-CPU=40
-GPU=
-SYNTHESIZERS=HMA1
-MODALITIES=multi-table
-CACHE=TRUE
