--- scripts/config/identity.conf
@@ -1,6 +1,0 @@
-NAME=Identity
-CPU=20
-GPU=
-SYNTHESIZERS=Identity
-MODALITIES=
-CACHE=FALSE
