--- scripts/config/independent.conf
@@ -1,6 +1,0 @@
-NAME=Independent
-CPU=10
-GPU=
-SYNTHESIZERS=Independent
-MODALITIES=single-table
-CACHE=TRUE
