--- scripts/config/medgan.conf
@@ -1,6 +1,0 @@
-NAME=MedGAN
-CPU=20
-GPU=2
-SYNTHESIZERS=MedGAN
-MODALITIES=single-table
-CACHE=TRUE
