--- scripts/config/par.conf
@@ -1,7 +1,0 @@
-NAME=PAR
-CPU=10
-GPU=2
-SYNTHESIZERS=PAR
-MODALITIES=timeseries
-CACHE=TRUE
-PARALLEL=TRUE
