--- scripts/config/privbn.conf
@@ -1,6 +1,0 @@
-NAME=PrivBN
-CPU=20
-GPU=
-SYNTHESIZERS=PrivBN
-MODALITIES=single-table
-CACHE=TRUE
