--- scripts/config/sdv.conf
@@ -1,6 +1,0 @@
-NAME=SDV
-CPU=40
-GPU=
-SYNTHESIZERS=SDV
-MODALITIES=multi-table
-CACHE=TRUE
