--- scripts/config/tablegan.conf
@@ -1,6 +1,0 @@
-NAME=TableGAN
-CPU=20
-GPU=2
-SYNTHESIZERS=TableGAN
-MODALITIES=single-table
-CACHE=TRUE
