--- scripts/config/tvae.conf
@@ -1,6 +1,0 @@
-NAME=TVAE
-CPU=20
-GPU=2
-SYNTHESIZERS=TVAE
-MODALITIES=single-table
-CACHE=TRUE
