--- scripts/config/uniform.conf
@@ -1,6 +1,0 @@
-NAME=Uniform
-CPU=10
-GPU=
-SYNTHESIZERS=Uniform
-MODALITIES=single-table
-CACHE=TRUE
