--- scripts/config/veegan.conf
@@ -1,6 +1,0 @@
-NAME=VEEGAN
-CPU=10
-GPU=2
-SYNTHESIZERS=VEEGAN
-MODALITIES=single-table
-CACHE=TRUE
