--- sdgym/constants.py
@@ -1,3 +1,0 @@
-CATEGORICAL = "categorical"
-CONTINUOUS = "continuous"
-ORDINAL = "ordinal"
