+++ sdgym/synthesizers/__init__.py
@@ -1,30 +1,20 @@
-from sdgym.synthesizers.clbn import CLBNSynthesizer
 from sdgym.synthesizers.generate import (
     SYNTHESIZER_MAPPING, create_multi_table_synthesizer, create_sdv_synthesizer_variant,
     create_sequential_synthesizer, create_single_table_synthesizer)
 from sdgym.synthesizers.identity import DataIdentity
 from sdgym.synthesizers.independent import IndependentSynthesizer
-from sdgym.synthesizers.medgan import MedGANSynthesizer
-from sdgym.synthesizers.privbn import PrivBNSynthesizer
 from sdgym.synthesizers.sdv import (
     CopulaGANSynthesizer, CTGANSynthesizer, FastMLPreset, GaussianCopulaSynthesizer,
     HMASynthesizer, PARSynthesizer, SDVRelationalSynthesizer, SDVTabularSynthesizer,
     TVAESynthesizer)
-from sdgym.synthesizers.tablegan import TableGANSynthesizer
 from sdgym.synthesizers.uniform import UniformSynthesizer
-from sdgym.synthesizers.veegan import VEEGANSynthesizer
 
 __all__ = (
-    'CLBNSynthesizer',
     'DataIdentity',
     'IndependentSynthesizer',
-    'MedGANSynthesizer',
-    'PrivBNSynthesizer',
-    'TableGANSynthesizer',
     'CTGANSynthesizer',
     'TVAESynthesizer',
     'UniformSynthesizer',
-    'VEEGANSynthesizer',
     'CopulaGANSynthesizer',
     'GaussianCopulaSynthesizer',
     'HMASynthesizer',
