+++ sdgym/synthesizers/base.py
@@ -1,10 +1,7 @@
 import abc
 import logging
 
-import pandas as pd
 import rdt
-
-from sdgym.errors import UnsupportedDataset
 
 LOGGER = logging.getLogger(__name__)
 
@@ -214,95 +211,3 @@ class MultiSingleTableBaselineSynthesizer(BaselineSynthesizer, abc.ABC):
             tables[table_name] = table[self.table_columns[table_name]]
 
         return tables
-
-
-class LegacySingleTableBaselineSynthesizer(SingleTableBaselineSynthesizer, abc.ABC):
-    """Single table baseline which passes ordinals and categoricals down.
-
-    This class exists here to support the legacy baselines which do not operate
-    on metadata and instead expect lists of categorical and ordinal columns.
-
-    NOTE: doesn't work with SDV 1.0.
-    """
-
-    MODALITIES = ('single-table', )
-
-    def _get_columns(self, real_data, table_metadata):
-        model_columns = []
-        categorical_columns = []
-        fields_meta = table_metadata.columns
-        for column in real_data.columns:
-            field_meta = fields_meta[column]
-            field_type = field_meta['type']
-            if field_type == 'id':
-                continue
-
-            index = len(model_columns)
-            if field_type == 'categorical':
-                categorical_columns.append(index)
-
-            model_columns.append(column)
-
-        return model_columns, categorical_columns
-
-    def get_trained_synthesizer(self, data, metadata):
-        """Get the trained synthesizer.
-
-        Args:
-            data (dict):
-                A dict mapping table name to table data.
-            metadata (sdv.Metadata):
-                The multi-table metadata.
-
-        Returns:
-            dict:
-                A mapping of table name to synthesizers.
-        """
-        self.columns, self.categoricals = self._get_columns(data, metadata)
-        data = data[self.columns]
-
-        if self.categoricals:
-            self.ht = rdt.HyperTransformer(default_data_type_transformers={
-                'categorical': 'LabelEncodingTransformer',
-            })
-            self.ht.fit(data.iloc[:, self.categoricals])
-            model_data = self.ht.transform(data)
-        else:
-            model_data = data
-
-        self.model_columns = model_data.columns
-
-        supported = set(model_data.select_dtypes(('number', 'bool')).columns)
-        unsupported = set(model_data.columns) - supported
-        if unsupported:
-            unsupported_dtypes = model_data[unsupported].dtypes.unique().tolist()
-            raise UnsupportedDataset(f'Unsupported dtypes {unsupported_dtypes}')
-
-        nulls = model_data.isnull().any()
-        if nulls.any():
-            unsupported_columns = nulls[nulls].index.tolist()
-            raise UnsupportedDataset(f'Null values found in columns {unsupported_columns}')
-
-        LOGGER.info("Fitting %s", self.__class__.__name__)
-        self.fit(model_data.to_numpy(), self.categoricals, ())
-
-    def sample_from_synthesizer(self, synthesizer, n_samples):
-        """Sample from the given synthesizers.
-
-        Args:
-            synthesizer:
-                The table synthesizer.
-            n_samples (int):
-                The number of samples.
-
-        Returns:
-            dict:
-                A mapping of table name to sampled table data.
-        """
-        sampled_data = self.sample(n_samples)
-        sampled_data = pd.DataFrame(sampled_data, columns=self.model_columns)
-
-        if self.categoricals:
-            sampled_data = self.ht.reverse_transform(sampled_data)
-
-        return sampled_data
